fix: pagination reviews and CircularProgress in loading document
